+++ src/lib/Components/Consignments/Screens/Welcome.tsx
@@ -38,23 +38,23 @@ const HorizontalLayout = styled.View`
   flex: 1;
   flex-direction: row;
   margin-left: ${isPad ? 0 : 20};
-  margin-top: 50;
+  margin-top: 40;
   margin-bottom: ${isPad ? 0 : 50};
 `
 
 const Title = styled.Text`
   text-align: center;
-  font-size: ${isPad ? 46 : 16};
-  line-height: ${isPad ? 58 : 32};
+  font-size: ${isPad ? 20 : 16};
+  line-height: ${isPad ? 24 : 32};
   width: ${isPad ? 760 : 300};
-  margin-top: ${isPad ? 80 : 35};
+  margin-top: ${isPad ? 0 : 20};
   font-family: ${Fonts.AvantGardeRegular};
   align-self: center;
   color: white;
 `
 
 const SmallHeadline = styled(LargeHeadline)`
-  font-size: ${isPad ? 34 : 20};
+  font-size: ${isPad ? 30 : 20};
   color: white;
   ${isPad ? "width: 540" : "max-width: 280"};
   min-height: 60;
@@ -69,6 +69,7 @@ const Icon = styled(Image)`
 const ButtonsView = styled.View`
   flex: 1;
   align-items: center;
+  margin-top: 30;
 `
 
 const List = isPad ? Listpad : Listphone